refactor: openReconnectWalletModal
